Refactor Board further. Cleanup code
