+++ src/pages/PageMain.js
@@ -1,10 +1,8 @@
-import Sidebar from "../components/Sidebar/Sidebar";
 import ProfileCard from "../components/ProfileCard";
 
 function PageMain() {
   return (
     <div className="planner-backg">
-      {/* <Sidebar /> */}
       <ProfileCard />
     </div>
   );
