+++ src/pages/Progress.js
@@ -1,10 +1,8 @@
 import React from "react";
-import Sidebar from "../components/Sidebar/Sidebar";
 
 function Progress() {
   return (
     <>
-      <Sidebar />
       <div className="temp-text">
         <h1>In Progress</h1>
       </div>
